fixed greedy meshing, started implementing world grid

+++ src/engine/engine.cpp
@@ -6,10 +6,10 @@ using namespace ve001;
 Engine::Engine(Vec3i32 chunk_size) : _engine_context(EngineContext{
     .shader_repo = {},
     .chunk_size = chunk_size,
-    .chunk_size_1D = static_cast<i32>(chunk_size[0] * chunk_size[1] * chunk_size[2]),
-    .chunk_voxel_data_size  = static_cast<i32>(chunk_size[0] * chunk_size[1] * chunk_size[2] * sizeof(u16)),
-    .chunk_max_mesh_size    = static_cast<i32>(chunk_size[0] * chunk_size[1] * chunk_size[2] * sizeof(Vertex) * (36/2)),
-    .chunk_max_submesh_size = static_cast<i32>((chunk_size[0] * chunk_size[1] * chunk_size[2] * sizeof(Vertex) * (36/2))/6),
+    .chunk_size_1D = static_cast<u64>(chunk_size[0]) * static_cast<u64>(chunk_size[1]) * static_cast<u64>(chunk_size[2]),
+    .chunk_voxel_data_size  = static_cast<u64>(chunk_size[0]) * static_cast<u64>(chunk_size[1]) * static_cast<u64>(chunk_size[2]) * sizeof(u16),
+    .chunk_max_mesh_size    = static_cast<u64>(chunk_size[0]) * static_cast<u64>(chunk_size[1]) * static_cast<u64>(chunk_size[2]) * sizeof(Vertex) * static_cast<u64>(36/2),
+    .chunk_max_submesh_size = static_cast<u64>(chunk_size[0]) * static_cast<u64>(chunk_size[1]) * static_cast<u64>(chunk_size[2]) * sizeof(Vertex) * static_cast<u64>((36/2)/6),
     .meshing_axis_progress_step = 32
 })
 {}
