+++ src/engine/engine_context.h
@@ -14,13 +14,13 @@ struct EngineContext {
     /// @brief chunk size/resolution 
     vmath::Vec3i32 chunk_size;
     /// @brief chunk size/resolution in 1D
-    vmath::i32 chunk_size_1D;
-    /// @brief chunk size in voxel context (single voxel)
-    vmath::i32 chunk_voxel_data_size;
+    vmath::u64 chunk_size_1D;
+    /// @brief chunk size in voxel context (single voxel chunk)
+    vmath::u64 chunk_voxel_data_size;
     /// @brief maximum possible mesh size of a single chunk after greedy meshing
-    vmath::i32 chunk_max_mesh_size;
+    vmath::u64 chunk_max_mesh_size;
     /// @brief maximum possible mesh size of a single chunk's side after greedy meshing
-    vmath::i32 chunk_max_submesh_size;
+    vmath::u64 chunk_max_submesh_size;
     /// @brief it maps to local_size_x attribute in greedy meshing compute shader
     vmath::i32 meshing_axis_progress_step;
 };
